+++ hdf_service/distributed_audio/hdi_service/audio/v1_0/src/audio_render_lowlatency_impl.cpp
@@ -369,8 +369,8 @@ int32_t AudioRenderLowLatencyImpl::GetExtraParams(std::string &keyValueList)
 int32_t AudioRenderLowLatencyImpl::ReqMmapBuffer(int32_t reqSize, AudioMmapBufferDescriptor &desc)
 {
     DHLOGI("Request mmap buffer.");
-    int32_t minSize = CalculateSampleNum(devAttrs_.sampleRate, minTimeInterval_);
-    int32_t maxSize = CalculateSampleNum(devAttrs_.sampleRate, maxTimeInterval_);
+    uint32_t minSize = CalculateSampleNum(devAttrs_.sampleRate, minTimeInterval_);
+    uint32_t maxSize = CalculateSampleNum(devAttrs_.sampleRate, maxTimeInterval_);
     int32_t realSize = reqSize;
     if (reqSize < minSize) {
         realSize = minSize;
@@ -379,7 +379,7 @@ int32_t AudioRenderLowLatencyImpl::ReqMmapBuffer(int32_t reqSize, AudioMmapBuffe
     }
     DHLOGI("ReqMmap buffer realsize : %d, minsize: %d, maxsize:%d.", realSize, minSize, maxSize);
     desc.totalBufferFrames = realSize;
-    ashmemLength_ = realSize * devAttrs_.channelCount * devAttrs_.format;
+    ashmemLength_ = static_cast<int32_t>(realSize * devAttrs_.channelCount * devAttrs_.format);
     DHLOGI("Init ashmem real sample size : %d, length: %d.", realSize, ashmemLength_);
     int32_t ret = InitAshmem(ashmemLength_);
     if (ret != HDF_SUCCESS) {
@@ -387,7 +387,7 @@ int32_t AudioRenderLowLatencyImpl::ReqMmapBuffer(int32_t reqSize, AudioMmapBuffe
         return HDF_FAILURE;
     }
     desc.memoryFd = fd_;
-    desc.transferFrameSize = CalculateSampleNum(devAttrs_.sampleRate, timeInterval_);
+    desc.transferFrameSize = static_cast<int32_t>(CalculateSampleNum(devAttrs_.sampleRate, timeInterval_));
     lengthPerTrans_ = desc.transferFrameSize * devAttrs_.channelCount * devAttrs_.format;
     desc.isShareable = false;
     ret = audioExtCallback_->RefreshAshmemInfo(adapterName_, devDesc_.pins, fd_, ashmemLength_, lengthPerTrans_);
